TP REVAMP :: TP calculations have been overhauled and are now accurate. Subtle Blow and Store TP are also working fully. See details.
Details : TP/hit now depends on the delay of the weapon, as on retail. Further, Store TP (from traits and gear) and Subtle Blow (from traits and gear) will affect the TP granted / hit and the TP the monster receives per hit. There were NO entries for Store TP in the modlist table, so I added a ton (but crucially not all!). Since there was a major change to the TP format (int16 -> float) there MAY BE PROBLEMS WHICH I HAVE NOT BEEN ABLE TO TEST. Please report any TP related problems! Finally, the monster's TP calculations (on when to use TP) has been adjusted slightly. You won't see any difference now, but you will when mob tp moves are added ;3

+++ src/map/battleentity.cpp
@@ -132,13 +132,12 @@ int32 CBattleEntity::GetMaxMP()
 
 /************************************************************************
 *																		*
-*  Изменяем количество TP сущности      								*
-*																		*
+*  Изменяем количество TP сущности										*
 ************************************************************************/
 
-uint16 CBattleEntity::addTP(int16 tp)
+uint16 CBattleEntity::addTP(float tp)
 {	
-	int16 cap = cap_value(health.tp + tp,0,300);
+	float cap = cap_value(health.tp + tp,0,300);
 	tp = health.tp - cap;
 	health.tp = cap;
 	return abs(tp);
